+++ Runtime/DataStream.cs
@@ -1543,6 +1543,18 @@ namespace Unity.Networking.Transport
             return (ushort)length;
         }
 
+        /// <summary>
+        /// Get a pointer to the stream's data. Note that the pointer always points at the
+        /// beginning of the data, no matter how much was read from the stream.
+        /// </summary>
+        /// <remarks>Performs a job safety check for read-only access.</remarks>
+        /// <returns>A pointer to the stream's data.</returns>
+        public unsafe void* GetUnsafeReadOnlyPtr()
+        {
+            CheckRead();
+            return m_bufferPtr;
+        }
+
         [Conditional("ENABLE_UNITY_COLLECTIONS_CHECKS")]
         void CheckRead()
         {